quantbridgev1/requirements.txt
PyYAML>=6.0
ctrader-open-api>=0.9.2
service-identity>=24.1.0
